activities is now aligned in rows
