+++ assets/styles/activitiesStyle.css
@@ -1,0 +1,5 @@
+#upcomingActivities{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
